requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
feedparser==6.0.10
beautifulsoup4==4.12.2
aiohttp==3.9.1
SQLAlchemy==2.0.23
pytz==2023.3.post1
httpx==0.25.2
